Refactor D3D Feature Processor class code
* Moved D3D Feature Processor code from original scripts to node
* Misc changes to processor class
* Added class import to top level processor
* Code has been tested. Some issue with angle rotations.
